# fix(ot): improve ideal functionality performance